Remove two podcasts
It's with a heavy heart.. I must Marie Kondo these podcasts..

Kermode and Mayo is fun, I like the idea of a weekly film podcast, starting during a pandemic perhpas not the best idea.. but really this is a recorded radio show, and radio is made for people in cars or to be left blaring at the back of an art studio whilst you wander around. It's too long for a podcast and too regular. I'd much prefer the 20-45 min version weekly

Oh Heavyweight, I don't know, I feel bad about removing this one but I'm just not enjoying it as much as I used to..

I feel like the amazing scenarios of the first season have become more and more strained as time has gone on.