Remove listening from navigation
Turns out in 8d0d0dd when i removed podcast.yml it was being used for
this page.

But this page is not doing a great deal. I think I am going to remove
this for now and have a rethink of what I want a podcast list to do and
how much time I want to put into maintaining it.

For now I can still host my OPML file if anyone wants to nick my
subscriptions.