Got a new job, time to update
It may be asked, is it wise to reveal this much information. I think so.
We work in the open[1], so much of this a canny github sleuth could
discover anyway. And we are encouraged to talk about our work too.

So... if you've managed to find this commit message, or decide to ever
look through the Git Blame on my website. This is who I am and what I
do. Please use my contact details to get in touch and ask me anything
you'd like.

I'll answer what I can and find the right person to talk to otherwise.

[1]: https://gds.blog.gov.uk/2017/09/04/the-benefits-of-coding-in-the-open/