Remove weeknotes from Nav
These are really old, I'm not really proud of the writing on any of
this site, and would like to start making a push in the new year to
do more (likely tech focused posts). These aren't helping I think.

I'll keep the URLs in place but remove the nav for now