Add review_url to decap CMS book object
It's been a while since I wrote any reviews, and I have an intention
to try and do more. I've enabled jekyll to zip together URLs and books
into nice HTML by manually providing the review URL on the book.

I never created a field for this, so I'm doing so now.