Refactor of wikidata.rb
This script uses the SPARQL endpoint to load books from frontmatter
and output a todo list of entries I'll want to create.

I've moved shared concerns off to the Helper, added a retry, and
accounted for timeout concerns whilst being more respectful of the
API usage.

We write the results to a scraped_books.json