Overhaul the check_wikidata_book_entries script
Returns a todo list of Mandatory and other fields I can use to target
work.

Initial focus is on Works, Editions is somewhat broken, but i should
be able to repuprose a lot of this.

key difference is that I want to do this:

```ruby
         if instance_of == "Q106833"
           {
             item["id"] => {
               "Missing Mandatory" => {
                 "change instance_of" => {
                   "is_currently" => "Q106833",
                   "should_be" => "Q122731938"
                 }
               }
             }
           }
```

That is, I want to rename a bunch of existing audiobooks to audiobook
editions